Se sube arreglo de los importes en Venta y Cobro.

+++ Magasys/AdminDashboard/CobroCliente.aspx.cs
@@ -110,7 +110,7 @@ namespace PL.AdminDashboard
 
                         lstVentaListado.Add(oVentaListado);
                         var loTotal = oVentaListado.TOTAL.Replace('$', ' ').Trim();
-                        loMontoTotal += Convert.ToInt32(loTotal);
+                        loMontoTotal += Convert.ToDouble(loTotal);
                         break;
                     }
                 }
